Added ability to create box bodies. Created a Box<Model class to render a simple box. When added physics bodies you must specify the shape of the body.
mask was supposed to be an array not a number

render_radius now has an option to generate a display list.

plane bodies can be instantied with a distance value.

leaveing BULLET commented out in the networking code was giving some warnings about Player::BULLET coming from an outside scope.

Created a wall of boxes in the main room with the other players.